backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pymongo==4.6.0
redis==5.0.1
newsapi-python==0.2.7
praw==7.7.1
nltk==3.8.1
spacy==3.7.2
feedparser==6.0.10
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
schedule==1.2.1
aiohttp==3.9.1 